LayoutTests/imported/w3c/web-platform-tests/scroll-animations/css/scroll-timeline-nearest-with-absolute-positioned-element-expected.txt


FAIL Resolving scroll(nearest) for an absolutely positioned element assert_equals: expected "100px" but got "300px"
